restructure app and create obsstate module
