+++ scripts/templates/ovro-lwa.ico
