+++ scripts/templates/settings.html
