--- tmp.sdf
@@ -1,23 +1,0 @@
-PI_ID            162
-PI_NAME          claw
-
-PROJECT_ID       0
-SESSION_ID       146
-SESSION_MODE     FAST
-CONFIG_FILE      /home/pipeline/proj/lwa-shell/mnc_python/config/lwa_config_calim.yaml
-
-OBS_ID          1
-OBS_TARGET      Cygnus A
-OBS_START_MJD   60278
-OBS_START_MPM   0
-OBS_START       UTC 2023 11 30 00:00:00.000
-OBS_DUR         60000
-OBS_INT_TIME    None
-OBS_DUR+        00:01:00
-OBS_FREQ1       1161394218
-OBS_FREQ1+      53.000000009 MHz
-OBS_FREQ2       1599656187
-OBS_FREQ2+      73.000000010 MHz
-OBS_BW          7
-OBS_BW+         19.600 MHz
-
